update(driver/bpf): move map definition struct to maps.h
