+++ userspace/libscap/scap_bpf.c
@@ -40,6 +40,7 @@ limitations under the License.
 #include "scap_bpf.h"
 #include "driver_config.h"
 #include "../../driver/bpf/types.h"
+#include "../../driver/bpf/maps.h"
 #include "compat/misc.h"
 #include "compat/bpf.h"
